server/requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
cfscrape==2.0.7
chardet==3.0.4
Click==7.0
Django==2.2.3
djangorestframework==3.10.1
idna==2.8
pytz==2019.1
requests==2.22.0
selenium==3.141.0
soupsieve==1.9.2
sqlparse==0.3.0
urllib3==1.25.3